+++ src/google_music_scripts/commands/sync.py
@@ -11,7 +11,8 @@ from google_music_scripts.__about__ import __title__, __version__
 from google_music_scripts.cli import (
 	CustomPath,
 	default_to_cwd,
-	parse_filters
+	parse_filters,
+	split_album_art_paths
 )
 from google_music_scripts.config import configure_logging
 from google_music_scripts.core import (
@@ -240,6 +241,11 @@ def sync_down(
 	callback=parse_filters,
 	help="Metadata filters."
 )
+@click.option(
+	'--album-art',
+	callback=split_album_art_paths,
+	help="Comma-separated list of album art filepaths.\nCan be relative filenames and/or absolute filepaths."
+)
 @click.argument(
 	'input-paths',
 	nargs=-1,
@@ -257,6 +263,7 @@ def sync_up(
 	max_depth,
 	delete_on_success,
 	filters,
+	album_art,
 	input_paths
 ):
 	"""Sync songs to a Google Music library."""
@@ -299,6 +306,7 @@ def sync_up(
 		upload_songs(
 			mm,
 			to_upload,
+			album_art=album_art,
 			delete_on_success=delete_on_success
 		)